+++ src/init.c
@@ -125,9 +125,8 @@ extern SEXP vctrs_fast_c(SEXP, SEXP);
 extern SEXP vctrs_data_frame(SEXP, SEXP, SEXP);
 extern SEXP vctrs_df_list(SEXP, SEXP, SEXP);
 extern SEXP vctrs_identify_runs(SEXP);
-extern SEXP vctrs_slice_complete(SEXP);
-extern SEXP vctrs_detect_complete(SEXP);
-extern SEXP vctrs_locate_complete(SEXP);
+extern SEXP vctrs_df_slice_complete(SEXP);
+extern SEXP vctrs_df_locate_complete(SEXP);
 extern SEXP vctrs_df_detect_complete(SEXP);
 
 
@@ -271,9 +270,8 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_data_frame",                 (DL_FUNC) &vctrs_data_frame, 3},
   {"vctrs_df_list",                    (DL_FUNC) &vctrs_df_list, 3},
   {"vctrs_identify_runs",              (DL_FUNC) &vctrs_identify_runs, 1},
-  {"vctrs_slice_complete",             (DL_FUNC) &vctrs_slice_complete, 1},
-  {"vctrs_locate_complete",            (DL_FUNC) &vctrs_locate_complete, 1},
-  {"vctrs_detect_complete",            (DL_FUNC) &vctrs_detect_complete, 1},
+  {"vctrs_df_slice_complete",          (DL_FUNC) &vctrs_df_slice_complete, 1},
+  {"vctrs_df_locate_complete",         (DL_FUNC) &vctrs_df_locate_complete, 1},
   {"vctrs_df_detect_complete",         (DL_FUNC) &vctrs_df_detect_complete, 1},
   {NULL, NULL, 0}
 };